Update baseOptions to use a RetryPolicy that tries up to 3 times.
